fixed NullPointerException when client connects to non-existing scope
git-svn-id: http://red5.googlecode.com/svn/java/server/trunk@870 1b6495e4-3631-0410-8e05-8f51eee8b9cc

+++ src/org/red5/server/net/rtmp/RTMPConnection.java
@@ -220,14 +220,16 @@ public abstract class RTMPConnection extends BaseConnection
 	}
 	
 	public void close(){
-		IStreamService streamService = (IStreamService) getScopeService(scope, IStreamService.STREAM_SERVICE, StreamService.class);		
-		synchronized (streams) {
-			for(int i=0; i<streams.length; i++){
-				IStream stream = streams[i];
-				if(stream != null) {
-					log.debug("Closing stream: "+ stream.getStreamId());
-					streamService.deleteStream(this, stream.getStreamId());
-					streams[i] = null;
+		IStreamService streamService = (IStreamService) getScopeService(scope, IStreamService.STREAM_SERVICE, StreamService.class);
+		if (streamService != null) {
+			synchronized (streams) {
+				for(int i=0; i<streams.length; i++){
+					IStream stream = streams[i];
+					if(stream != null) {
+						log.debug("Closing stream: "+ stream.getStreamId());
+						streamService.deleteStream(this, stream.getStreamId());
+						streams[i] = null;
+					}
 				}
 			}
 		}
